BREAKING ContourModel (see below) + Fixed review remarks.
+ fixed gcc compiler errors
+ Refactored ContourModel interface to use mitk::TimeStepType
  instead of int/unsigned int for time steps. Code might break that
  currently explictly casts timesteps (that where provided from code
  using the correct type (TimeStepType)) to int/unsigned int. Correct
  or remove (where appropriate) the cast.

+++ Modules/Segmentation/Interactions/mitkFeedbackContourTool.cpp
@@ -169,7 +169,7 @@ void mitk::FeedbackContourTool::AddVertexToFeedbackContour(const Point3D& point,
     return;
   }
 
-  m_FeedbackContour->AddVertex(point,static_cast<int>(feedbackTimeStep));
+  m_FeedbackContour->AddVertex(point, feedbackTimeStep);
 }
